Update README.md
My cousin is about to relocate to ST AUSTIN which is in the UK and she intends to continue doing her tailoring business close to her new place so she reached out to me saying As a tailor , I want to rent a places so I could use it as  my shop and  I want to rent a place in a location that is safe from theft so that her property, equipments and customer’s belongings are safe. So i knew the acceptance criteria here was she does not want a location where there are high cases of theft. So i had to search for data i could  use for this and luckily for me i saw a data set showing all crimes in StAustin ranging from rape,assault, theft and all other sort of crimes. So I had to filter my data set to get the information i needed which was primary_type ‘theft’ and there i was able to find the location in StAustin with the highest theft cases which was 15TH StAustin both in the year 2015 and 2016. Therefore, I advised my cousin she should not look for a place in 15th StAustin although the data set ends at the year 2016 making it old and outdated but it showed me a great insight